Add files via upload
This log_attack is a special feature of the basher robot. This file contains a special data request for the skull basher skill, which allows the basher robot to receive additional damage and a 2-second stun effect. This data must be injected into the JSON file. It can then be read by log_death for each hit event.